[process32] ASSERT curr != NULL in switch_to_task()
NOTE: ASSERT(curr != NULL) also in is_fpu_enabled_for_task().

+++ kernel/process.c
@@ -669,7 +669,7 @@ sptr sys_fork(void)
       return -EAGAIN;
    }
 
-   task_info *child = allocate_new_process(curr, (u16) pid);
+   task_info *child = allocate_new_process(curr, pid);
 
    if (!child) {
       enable_preemption();
